+++ scheduler/handle_update.go
@@ -73,7 +73,6 @@ func (e *Scheduler) HandleUpdate(event *mesosproto.Event) {
 		e.Redis.DelRedisKey(task.TaskName + ":" + task.TaskID)
 		task.TaskID = e.API.IncreaseTaskCount(task.TaskID)
 		task.State = ""
-		break
 	case mesosproto.TaskState_TASK_LOST:
 		if task.TaskID == "" {
 			e.Mesos.Call(msg)
@@ -81,9 +80,16 @@ func (e *Scheduler) HandleUpdate(event *mesosproto.Event) {
 		}
 		logrus.WithField("func", "scheduler.HandleUpdate").Warn("Task State: " + task.State + " " + task.TaskID + " (" + task.TaskName + ")")
 		e.Redis.DelRedisKey(task.TaskName + ":" + task.TaskID)
-		e.Mesos.ForceSuppressFramework()
-		e.Mesos.Call(msg)
-		return
+
+		// Remove task if lost as a default behaviour.. This can be overriden using ENV variable "TASK_LOST_REMOVE_TASK" which will restart the task
+		if e.Config.TaskLostRemovesTask {
+			e.Mesos.ForceSuppressFramework()
+			e.Mesos.Call(msg)
+		} else {
+			task.TaskID = e.API.IncreaseTaskCount(task.TaskID)
+			task.State = ""
+		}
+
 	case mesosproto.TaskState_TASK_RUNNING:
 		if !e.Mesos.IsSuppress {
 			logrus.WithField("func", "scheduler.HandleUpdate").Info("Task State: " + task.State + " " + task.TaskID + " (" + task.TaskName + ")")
@@ -92,8 +98,8 @@ func (e *Scheduler) HandleUpdate(event *mesosproto.Event) {
 		task.MesosAgent = e.Mesos.GetAgentInfo(update.Status.GetAgentId().GetValue())
 		task.NetworkInfo = e.Mesos.GetNetworkInfo(task.TaskID)
 		task.Agent = update.Status.GetAgentId().GetValue()
-
-		e.Mesos.SuppressFramework()
+	default:
+		logrus.WithField("func", "scheduler.HandleUpdate").Warn("Task State: " + task.State + " " + task.TaskID + " (" + task.TaskName + "). State not handled, no action has been taken")
 	}
 
 	// save the new state